Attempt to use -O3 and -flto if available on gcc
RE2 compiles with -O3 by default, so we override perl's default
OPTIMIZE level with -O3 to match. Setting this is avoided if OPTIMIZE=
is specified on the Makefile.PL command line though.

-O3 seems to give about a 6% speedup in my benchmarks. I don't have a
box running gcc 4.5 that isn't a VM, so I don't know about -flto for
sure, it doesn't seem to slow things down.