Support fallback to Perl RE under >= 5.17.1
Perl_re_compile now takes the RE engine in account, previously it would
always use the Perl regexp engine, which the fallback code took
advantage of.

It doesn't seem like there's a way of using a public published API to
choose to use the Perl engine anymore, but Perl_re_op_compile seems to
work.

(This might break on Windows and other platforms that are strict about
linking.)